refactor(aq3): isolate queue bootstrap configuration
